refatorando...
