+++ .gitignore
@@ -1,1 +1,2 @@
 record.txt
+__pycache__/
