+++ view.py
@@ -1,0 +1,9 @@
+import os
+
+def limpa_tela():
+    os.system('cls||clear')
+
+def titulo():
+    print('-'*50)
+    print(f'{"Palavra Oculta":^50}')
+    print('-'*50)
